Integrate merge operator with @felipeblazing implementation
