--- src/include/planner/sirius_plan_utils.hpp
@@ -1,35 +1,0 @@
-/*
- * Copyright 2025, Sirius Contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include "duckdb/common/vector.hpp"
-#include "duckdb/planner/bound_result_modifier.hpp"
-
-namespace sirius::planner {
-
-inline duckdb::vector<duckdb::BoundOrderByNode> copy_order_nodes(
-  const duckdb::vector<duckdb::BoundOrderByNode>& orders)
-{
-  duckdb::vector<duckdb::BoundOrderByNode> copies;
-  copies.reserve(orders.size());
-  for (const auto& order : orders) {
-    copies.push_back(order.Copy());
-  }
-  return copies;
-}
-
-}  // namespace sirius::planner
