+++ src/op/sirius_physical_ungrouped_aggregate.cpp
@@ -22,6 +22,7 @@
 #include "duckdb/planner/expression/bound_aggregate_expression.hpp"
 #include "duckdb/planner/expression/bound_reference_expression.hpp"
 #include "op/merge/gpu_merge_impl.hpp"
+#include "op/sirius_physical_ungrouped_aggregate_merge.hpp"
 
 #include <cudf/column/column_factories.hpp>
 #include <cudf/fixed_point/fixed_point.hpp>
@@ -376,14 +377,43 @@ std::vector<std::shared_ptr<cucascade::data_batch>> sirius_physical_ungrouped_ag
   return outputs;
 }
 
+// Helper to deep copy Expression vector (same as in grouped_aggregate)
+static duckdb::vector<duckdb::unique_ptr<duckdb::Expression>> copy_expressions(
+  const duckdb::vector<duckdb::unique_ptr<duckdb::Expression>>& src)
+{
+  duckdb::vector<duckdb::unique_ptr<duckdb::Expression>> result;
+  result.reserve(src.size());
+  for (const auto& expr : src) {
+    result.push_back(expr->Copy());
+  }
+  return result;
+}
+
+sirius_physical_ungrouped_aggregate_merge::sirius_physical_ungrouped_aggregate_merge(
+  sirius_physical_ungrouped_aggregate* ungrouped_aggregate)
+  : sirius_physical_ungrouped_aggregate_merge(
+      ungrouped_aggregate->types,                         // copied by value
+      copy_expressions(ungrouped_aggregate->aggregates),  // deep copy
+      ungrouped_aggregate->estimated_cardinality,
+      duckdb::TupleDataValidityType::CAN_HAVE_NULL_VALUES)  // default - not stored in source
+{
+  child_op = ungrouped_aggregate;
+}
+
 sirius_physical_ungrouped_aggregate_merge::sirius_physical_ungrouped_aggregate_merge(
   duckdb::vector<duckdb::LogicalType> types,
   duckdb::vector<duckdb::unique_ptr<duckdb::Expression>> expressions,
-  duckdb::idx_t estimated_cardinality)
+  duckdb::idx_t estimated_cardinality,
+  duckdb::TupleDataValidityType distinct_validity)
   : sirius_physical_operator(
-      duckdb::PhysicalOperatorType::EXTENSION, std::move(types), estimated_cardinality),
+      SiriusPhysicalOperatorType::MERGE_AGGREGATE, std::move(types), estimated_cardinality),
     aggregates(std::move(expressions))
 {
+  distinct_collection_info = duckdb::DistinctAggregateCollectionInfo::Create(aggregates);
+  // aggregation_result       = duckdb::make_shared_ptr<GPUIntermediateRelation>(aggregates.size());
+  if (!distinct_collection_info) { return; }
+  distinct_data =
+    duckdb::make_uniq<duckdb::DistinctAggregateData>(*distinct_collection_info, distinct_validity);
 }
 
 std::vector<std::shared_ptr<cucascade::data_batch>>
